Move the Qt geometry opening GUI logic to qtcad
There shouldn't be any reason this logic can't be reused by other
applications - refactor it into libqtcad.  Right now we only have
dialogues for 3dm and STEP, but it shouldn't be too hard to add more.
At a minimum need to match Archer's dialogues.  In an ideal world this
would be some sort of generated list and dialogues based on runtime
info from gcv/libgcv, but that'll have to wait for both maturation of
gcv and the resources to figure out implementing such an approach.