Update compile tests for -Wstrict-prototypes
The bu_file realpath regression test was failing with clang 16.  Upon
investigation, the realpath detection proved to not be succeeding due to
the following error:

error: a function declaration without a prototype is deprecated in all
versions of C [-Werror,-Wstrict-prototypes]
int main() {(void)realpath; return 0;}
        ^
         void
1 error generated.

This is a more general issue with our tests that define functions
without arguments, and is also triggered by CMake's own
CheckStructHasMember.cmake test.  Apparently
(https://stackoverflow.com/a/76371638) this has to do with support
for void func() style functions being obsolete for a long time, and
the compilers are now adding warnings about them.

Looking in CMakeErrors.log for the warning message, we can identify
various cases in our code where tests were failing because of this
issue.  It's a little hard to check for by pattern searching, so tests
not run on this particular machine/configuration may still have issues
(and I've not checked at all to see if the third party build tests have
high enough strictness flags to hit this issue...)