Make the pkg_conn struct private
This hides the implementation details of the struct pkg_conn container
and adds accessor or get/set functions for the info that is accessed
outside of libpkg.

This is intended to be a preliminary to adding communication mechanisms
other than TCP/IP to libpkg's toolbox "under the hood" via libuv -
ideally, we'd want to put libuv fully underneath libpkg's API and let it
sort out things like named pipes vs. sockets.

The client codes of libpkg appear to expose some direct socket
manipulations, so that could have implications if we're not using
sockets for communication...