Fix executable permission on sh files
Looks like we lost executable settings in the 2023-02-02 year updating
commit.  Was manifesting in a distcheck failure when the check target's
invocation of benchmark clobber wasn't succeeding due to benchmark not
being executable, leaving behind files.  That calling of the benchmark
script didn't prefix with an explicit ${SH_EXEC} call, and thus relied
on benchmark being directly executable on its own.