Add a way to suppress the PNG use of 'd' suffix
Grr.  If build type gets set to Debug, instead of being left empty,
Linux is also throwing in the 'd' suffix on debug build libs from png.
Add a way to specify at configure time not to do that - it really messes
with our build logic.  Should probably try to upstream this or something
like it.