Remove odd_pathnames from distcheck-full set
Too problematic with 3rd party builds - we'll just keep it as a
stand-alone test when conditions are right to run it.  We may be able to
make it standard again if we fully separate src/other/ext into its own
repository and make it a separate build, but that's down the road.