Strip down libtermio code, eliminate public header
This doesn't quite get rid of libtermio completely, but it strips the
code down to the minimum in current use and makes it internal.  Next
step is to see if linenoise can be used with btclsh and the fb tools to
completely eliminate the need for the last termio functions (and maybe
even get those tools working on Windows).  However, as that change has
functional implications do this scrubbing first as a quick, lower risk
intermediate step.