fix validation issues fix prompting display a different way since just removing term doesn't
term is not valid in a para context, only in a varlist, but looks like we don't even really want that.  simplify the tagging and update the indent while we're in there, particularly for the three variants of gqa and check.  this ensures the prompts line up right with their output, and there's a space after the prompt (as they're all in literallayouts) for better readability.  passes build checks.