Create a space for self-contained code snippet examples
This initial script being added was created for a user that had
imported geometry using myriad methods resulting in regions with
less-than-ideal object names and incorrect IDs set.  The script fixed
those regions and is a simple general pattern for how to iterate over
a set of objects to perform an action not (yet) achievable with
"search -exec".

The intention here is to create a space for curating brief but useful
and easily understood code snippets that users might be able to adapt
and learn from.  As such, this script is succinctly documented with an
example how to run the script and what effect it has on a sample tree.
This is intended to be a pattern for future scripts that get added.
Scripts should be short enough (ideally less than 100 lines) to be
fully understood but longer than 1-liners typically provided via
manual pages.