Schedule big E command to be replaced by a drawing mode option to 'draw'
The "big E" command for evaluated wireframes is seldom used with modern
geometry, and conflicts with the 'e' man page for the draw command's
abbreviation on case insensitive filesystems.

Make the same capability available as a "mode" option to the draw
command (at the moment, the option is -m5) and schedule "E" for removal
as a top level command.