Add a drawing test to check view behavior
The idea of this test is to check that the view settings are stable
across dm events (such as resize) and bv_update calls.  Probably not
stressful enough as it stands, but it does at least check the resize
behavior of a dm, which is useful in and of itself.