spent few hours reviewing and rewriting rtwizard's manual page.
goal was to improve the readability, eliminate some of the technical detail and caveats, and make it more easily skimmed for information.  really pushed down (in some spots eliminated) all the distracting commentary about the implicit option support, now relegated to its own section at the end.  less emphasis on what tool underpins a given output style and speak more directly to the user-facing feature.  completely rewrote the intro to delineate the two operating modes.  mostly kept options and examples as-is.