Note a problem with current ORIGIN rpath logic
With the introduction of plugins and more complicated hierarchies, the
ORIGIN/../lib pattern isn't correct in all cases.  Shows up when testing
a tar.gz build without the install in its /usr/brlcad location - we're
not finding plugins correctly.

Need to think about how to handle this - probably need to process the
relative paths of the target objects and come up with the correct ORIGIN
path on a case by case basis.