Fix archer crash with "make arb8.s arb"
libtclcad/Archer implicit assumptions strike again. The libtclcad view
creation command assigns some custom structures to u_data, and the
default ged bview we were creating didn't have that.  The libtclcad
logic then got the default view, assumed something was there that
wasn't, and crashed.

For now we won't create the default view and will instead alter the
codes that had been assuming it to provide their own instead. I can see
arguments both ways for providing a default.  If we CAN provide a
default view with a ged_open, a whole lot of commands that would
otherwise have to bail unless/until the app provides a view will work
out of the box...  On the other hand, it may be reasonable to expect
applications to manage their own views - after all, that's what Archer
is doing.  Nor does qged rely on a ged provide view...