Rough in ged opendb/closedb callbacks, cleanup
The opendb command was repeating a lot of the work of the closedb
command - just have opendb call closedb first if needed.

As a first cut, give applications a chance to take action both
immediately before and immediately after the open and close operations
take place - it's conceivable they'll want to record info about the old
state, and they'll almost always want to respond to the new data from a
successful open.