Rework MGED f_opendb/f_closedb functions to use GED
The garbage_collect command exposed an interesting problem with how we
were managing the database open/close process.  The db_i struct
maintains a dbi_uses slot where applications can note their use of a
database instance, which will in turn instruct db_close to not
completely eliminate a dbip until the use count has been reduced to
zero.  This offers a form of protection from dbips becoming invalid
unexpectedly, and is used by MGED when defining Tcl interfaces to the
database.

However, this mechanism causes a problem when attempting to garbage
collect a database.  The current GC mechanism involves creating a new
version of the database and swapping it in for the old version.
However, doing so inevitably requires a full close operation on the
currently open database, and if the application has multiple uses of the
dbip registered libged lacks the information to properly close it.
While Linux will allow the file rename to succeed anyway (which is why
this issue wasn't caught in initial testing), Windows (correctly)
identifies the .g file as still being fopened and refuses to remove or
rename to overwrite it - the garbage collection itself succeeds, but we
can't stage the file to replace the old copy and re-open it without
breaking the dbi_uses rules.

Because libged itself can't know what the application needs, we add
callbacks to libged that the application can register in order to
execute necessary additional logic before and after the GED opendb and
closedb operations.  Thus far this is still only minimally tested on
Linux, so we need to confirm it truly solves the problem on Windows.  We
also need to confirm it correctly handles the various failure cases and
user interactions wrapped up with database I/O.