Add notes to update add_custom_command post-3.20
CMake in 3.20 added some generator expression support for OUTPUT in
add_custom_command.  Working around that limitation for multi-config
builds imposed some restrictions, and once we can require that version
(and confirm the expressions work successfully) there are some aspects
of our build logic that should be revisited for potential
simplification.  However, we'll have to bear in mind the limitation that
expressions there can't refer to targets...

https://cmake.org/cmake/help/latest/command/add_custom_command.html
https://gitlab.kitware.com/cmake/cmake/-/issues/21364