Have GED define both opendb/open and closedb/close
We need opendb and closedb to avoid collision with Tcl commands, but open and close should
also be options for environments that don't need to worry about collision with those names.
As long as GED supports both opendb/closedb and open/close, even if the latter are masked
by environment the former are still visible.  Conversely, if the enviornment doesn't mask
them open and close will do what is expected in the GED context.