Set up BU_DIR_CACHE env var to override the bu_dir cache location
Looking at possibility of unit testing cache behavior, and if we do that
we definitely DON'T want to target the real BU_DIR_CACHE location.
Beyond polluting the user cache with non-working files, that would be a
major potential issue for multiple BRL-CAD builds simultaneously doing
unit testing with the same cache tests.  Allowing for an env variable
gives us a way to target a local build directory in those cases instead.