Checkpoint early stages of libbu cache API
Since both the new drawing code and libbg's LoD logic are using the LMDB
key/value cache, it makes sense to see if we can consolidate the LMDB
logic common to both.  In principle the key/value store functionality
should be fully generic, so it seems like a logical fit for libbu.

I suspect my initial implementation of the LMDB calls wasn't optimal,
but some experimentation will be necessary since docs are a bit sparse
for these sorts of cases.  Also will want a function to return all keys
in the cache (initially for debugging, but it may have other uses too) -
I think this will need a slightly more advanced LMDB API call.  Look at
https://github.com/LMDB/lmdb/blob/mdb.master/libraries/liblmdb/mdb_dump.c
for some hints.

The idea will be to first set up unit tests of the cache and hammer it
with more easily validated inputs and outputs - both current uses of
this mechanism (particularly the LoD) tend to show success/failure
visually, unless the lookups outright fail.  (That's one reason I didn't
get adventuresome with the LMDB calls initially.)  However, if we can
get data out of the cache faster with parallel reads, it would be very
nice to support that use case.  Not certain if that's possible or not,
but worth checking.