Start moving BRL-CAD target dep lists to src/CMakeLists.txt
I'd really like to do better than the level 1,2,3 logic for enabling and
disabling parts of BRL-CAD for building, but the dependencies of various pieces
on each other complicate matters.  Should be able to learn from the bext work
and come up with something better, but will definitely need high level
information (i.e. before we start doing add_subdirectory calls) on the various
dependencies.

In order to avoid any problems with the high level dependency info and the
per-target listings getting out of sync, go with a DRY approach and define the
high level lists in such a way they can be used directly by the target
definitions as well.

Only did the basic libraries as an initial test, and this only defines the
lists - there isn't yet any new enable/disable logic.  This will be written
with an eye towards the new bext dependency approach, so third party build
enable/disable logic won't be a part of this - all such decisions are made in
the bext configure process.  However, if so desired, the eventual
add_subdirectory decisions for BRL-CAD components can also be based on
find_package results.

One important caveot - if no specific components are specified for building, we
need to add everything.  This will still be the default configuration, and the
standard build and distcheck logic needs to see everything.  The component
builds are intended for specific user use cases that neither need nor desire a
full BRL-CAD install.  It is intended to replace the BRLCAD_LEVEL* arrangement.