First step towards a bext directory list
There's a little bit of a chicken/egg problem with bext and brlcad -
only brlcad knows which 3rd party dependencies it needs, but only bext
can build them.  However, bext needs to be ready BEFORE the brlcad
configure process.  So if we're doing a minimal build, right now the
user has to know (or figure out) which parts of bext can be safely
turned off for a minimal subset.

One possible solution to this is to have BRL-CAD's logic localize the
find_package calls near the logic using them, and have those
find_package calls (which could all fail) print out the corresponding
bext directory that satisfies the requirement.  Essentially, it would
be a special type of configure whose job is simply to resolve the 3rd
party dependency needs and report them in a form that could be fed to
bext.  It's not clear yet how practical this will be - this commit just
prints out all the directories from find_bpackage calls.