Move libbv-specific functionality from bg to bv
Working with this logic, it's becoming clearer that libbg should be the
lower level library.  Shift the LoD and view-specific aspects of the
polygon logic to libbv.

Still need to move the plot3 logic - right now the libbg plotting
routines are just turned off while focusing on other changes, but that
won't fly long term.