Adjust view polygon logic to use planes, 3d pt
It was proving confusing to have view polygons store a complete view
copy - when to operate on or use the app view vs the polygon view, when
and whether to sync view states, etc.  Instead, have the polygons store
a plane.  Also, have apps supply 3D points in lieu of having the lower
level polygon logic try to handle the snapping logic.

This needs more shaking out, but we're as far as passing the regression
tests so it's a good time to checkpoint.