fix edcolors external opener on windows
Windows does a pretty good job of cleaning up temp files as soon as
they're closed. So, this method of opening, writing the current state,
closing, and then opening again to read is not reliable. Instead,
generate a name with bu_temp_file_name and open with w+ so the file is
guaranteed to stick around until it is removed with bu_file_delete