Rename OP_NMG_TESS to OP_TESS
This operation shouldn't be specific to NMG - there is a lot of logic
for the tessellation process that should be generic to the librt tree
structure and walking it.

Background: when testing alternative tessellation methods, I keep
writing quick, poor-man's tree walkers to fire off the experimental
boolean evaluation logic.  Inevitably, if I get far enough, I find I'm
not walking the tree the way the NMG facetization routines do.

Rather than keep doing that, I want to look at whether the non-NMG
aspects of the facetization tree walk can be made generic, so we can
swap in whatever evaluation method we want and have the CSG trees be
handled the same way.