nmg_booltree_evaluate -> rt_booltree_evaluate
Make a version of nmg_booltree_evaluate that abstract out the actual nmg
boolean evaluation into a callback function, in order to allow the
substitution of other evaluation methods in lieu of nmg_bool.

Hopefully, this will make for easier apples-to-apples comparisons with
more complex CSG tree hierarchies when evaluating alternative boolean
methods.