Enable a bundled manifold, given C++17 support
Boolean results seen on some of the example CSG targets are impressive,
so go ahead and get a bundled Manifold set up so we can do additional
testing more easily on various models and platforms.

Will need to think about how to handle non-manifold inputs if this
becomes a default option (facetize -r --MANIFOLD on the faa Generic Twin
ends up ditching most of the BoTs, as they're not manifold).  May want
to fall back to NMG if one of the inputs isn't a manifold, but I'm not
sure how much better NMG is for those cases - needs some
thought/testing.