Need alternative for MGED bu_log callback if on Windows
Since ged_exec calls are blocking for GUI updates, we try to leave the
bu_log callback that forcibly inserts text from other threads on if the
platform isn't Windows.

Deliberately using an #ifndef _WIN32 here, as we want to keep the
immediate feedback if we can get away with it for long running
subcommands like gqa and rt.  On non-Windows platforms it usually seems
to work.

The right answer here long term is to get ged_exec running in its own
thread, so we can keep the GUI refresh going "normally" and have the
text update calls come from there.  (That's the mged_pr_output call in
refresh() - right now it's fairly useless as refresh() is blocked, but
once we have ged_exec calls asynchronous it should start doing the job.)