Note changes in facetize speed and success rates
Refining what to report here a bit - this test is bot only output of
share/db/*.g with the filter:

export OBJECTS="\( -above -type region -or -type region \) -and \( -not -above -type half \)"

However, the most interesting cases for this particular change are the
complex ones where bool plays a major roll.  So these numbers may be
re-run with an additional filter such as -and \( -above -nnodes \>2 \)
to hone in on those more difficult cases.