Add brep repair subcommand to generate new edge curves
Probably shouldn't be announced to users just yet, but checkpoing progress so
far.  We hit a case (possibly caused by a flaw in one of our older importers)
where we have valid 2D trim curve data but bad 3D edge curves.  This command
uses one of the 2D trim curves and tries to generate new 3D edge curves.  Not
sure yet if it works, or if so how well.