Update numbers for just higher level conversions
Re-run with the following filter:
export OBJECTS="\( -above -type region -or -type region \) -and \( -not -above -type half \) -and \( -above -nnodes \>2 \) -and \( -depth \<3 \)

Translated, that selection looks for regions or assemblies with
trees not containing half spaces (which are problematic for
facetization), have at least one comb with two nodes in their
tree (to avoid trivial trees and test the boolean logic) and are at
depth one or two overall in the .g file (and thus likely candidates for
complex user facetization operations.)

While running these tests, discovered that (if I'm understanding this
logic correctly) the boolean tree walk for facetization is intended to
be (or at least, as implemented ends up being) tolerant of lower level
failures.  This can result in "successes" that are skipping individual
lower level components that don't work - this may be intended since it
does at least give the user partial results, but could also be a
surprise if pieces end up facetized oddly or are missing.

Perhaps we should consider an explicit option to facetize to be fault
tolerant and/or fail fast and hard if lower level pieces fail.