Backport a libpng fix from bext
Newer Ubuntu systems have a freetype that links in libpng as a shared
library, and using a default name for BRL-CAD's copy results in freetype
trying to use our file.  However, because we prefix our function names,
freetype can't find the symbols it needs.  (In a sense the symbol
prefixing is doing its job, as we want freetype to be using the system
.so it expects and was compiled for.)  To avoid this, we do the same
thing we do for zlib and adjust the library output name.