Teach libicv to generate a diff image
Not 100% sure we're properly matching pixdiff yet, but it's the right
general idea... Teach libicv to programmatically generate diff images,
so we can automatically write them out when drawing tests fail.

Usually, the first thing the dev wants to do is inspect the differences
between the control image and the generated image, so we go ahead and
provide that data as a PNG image under various failure conditions.