Begin reorg of facetize logic
facetize.cpp is getting unwieldy, and there's more to do to improve the
processing.  We want to merge the manifold and nmg methods (we should
simply try manifold first and then nmg if manifold doesn't work) and
attempt the fallback methods on individual primitives if their
individual nmg routines fail (rather than sampling whole large shapes
all the time, try to keep as close to a "standard" nmg bool process as
we can.)  We can still try the all-up CM/SPSR if individual shapes fail,
but hopefully that will become a rare situation.

We will also need, if possible, some means of automatically cleaning up
meshes that are "almost solid" but have small and localized topology
flaws - such meshes are often seen, and we don't want them to be the
reason we can't apply manifold and/or NMG boolean evaluation to the
tree.  Also need to special case half planes - for subtraction or
intersection we can substitute sufficiently large arbs, and they need to
be filtered out when unioned.

All this will require some adjustment to how facetize's pipeline works,
in addition to new support for operations like hole filling.  As a first
step, break logic into multiple files.  It's not working in this form as
yet - this is just a start.