Start working on allowing for a fallback tess clbk
If we wish to try alternatives to the standard librt tessellation
methods as part of the initial primitive tree walk, we'll need a way to
inject that logic into the walking and evaluation process.

We'll also need to be sure we can distinguish between an error, and a
deliberate decision by tessellate not to return a mesh.  In the latter
case we may want to accept that result and proceed, but we need to be
sure the return codes will let us distinguish correctly.  Will probably
have to audit the librt tessellation methods to see what they're doing
in such cases.