Use spheres and cylinders to fill in bot extrude
Not 100% sure how well this will work, but attempt to "fill in" the gaps
left by extruding bot face arb6 primitives straight out with cylinders
and rccs.  Subtracting a plate mode bot from this comb won't result in
an empty raytrace, since the result defines solid volumes extruding out
from the edges perpendicular to the plane of the faces, but arguably
that's OK as those volumes can be interpreted as being in the "solidity
depends on interrogation angle" ambiguities inherent to the plate mode
methodology.

In theory it may be possible to use facetize to evaluate this and
actually get an evaluated mesh, but initial experiments suggest that for
anything beyond fairly trivial surfaces the boolean eval will be
prohibitively expensive even if it succeeds.