Concentrate the NMG output specific logic.
If we're targeting an NMG output, we need to use the old-school libnmg
methods of evaluation (despite their lower success rate.)  Focus the NMG
code on that use case.  If we want to take an NMG output and generate a
bot from that for some reason, we should facetize the NMG output in BoT
mode.  If we're interested in triangles we're going to want to use the
newer logic - NMG mode is for the case where we specifically *don't*
want to step the data down to triangles.