Checkpoint an rt_comb_tess experiment
This sets up (more or less) what an rt_comb_tess callback for ft_tessellate
would look like, but also makes it clear why this hasn't been done previously -
due to the function signature of the RT_DECLARE_INTERFACE tess callback it
requires rt_db_internal to have awareness of at least dbip and (at the moment)
dp as well.  While this is not the first time I've wanted those, my
understanding is that conceptually the rt_db_internal entities are normally
"detached" from their source db_i and directory containers.  If they are
supposed to be prepared to persist beyond the survival or updating of their
sources, then the stashed pointers required by this approach may eventually, in
some usage scenario, become pointers to invalid data.

This is something of a dilemma - from an API cleanliness standpoint, an
rt_comb_tess function is the right place to put the "translate a comb object to
NMG" logic.  However, unless we clutter the rt_db_internal struct as we have
here, moving the logic also requires inputs not available in the callback
signature.  My current next step is to consider a change the
RTFUNCTAB_FUNC_TESS signature to pass in a dbip and dp as well.  The
import/export function signatures at least offer a precident for passing in a
dbip to one of the RT_DECLARE_INTERFACE functions, but none of them accepts a
struct directory.  Needs discussion before redoing all of the rt_*_tess
function signatures.