Start thinking about how to stream in bu_external
If temp files are going to be super slow, we want to be able to pipe data in
and out unless it's the rare case of some individual object so big we want to
keep its footprint in memory minimal.  We'll need logic on both ends for this,
but start by working on figuring out what the receiving end of ged_tessellate
should look like.  We'll want to preserve the option to use an input .g file as
well (for debugging if nothing else) so we'll also need to keep the .g file
version of the logic.

Note that this isn't tested yet, beyond compiling on Linux - one of the
drawbacks of the piping approach is we need both halves of the equation before
testing...