Back up to ecff08e99 - pipes are problematic
The more I research non-blocking pipes in a cross platform context, the worse
it gets. Time to take a step back and see if there's another way to come at
this.

Right now, each unique instance of each primitive in the tree walk is getting
written to its own temp file to be facetized, in order to dupliate the original
tree walk behavior of calling ft_tessellate on the instanced data of the
primitive.  The main advantage of this is that we don't have to communicate
anything other than the tree walk's geometry to the subprocess - matrix
positioning of the input data is handled by the tree walk. However, we end up
creating as many files as we have instances - which can be a LOT of small file
creation and open/close operations.

An alternative would be to make one copy of the .g file as the working copy,
and then do all the non-translated dp object facetizations once in that copy.
If we wanted to do parallel processing of multiple objects, Victor had the idea
to pass multiple objects to ged_tessellate, let it do the parallelization to
minimize the processes needed per object, and in the event of failure recover
by individually re-processing the group that failed to identify which object
was the "poison pill."  The latter involves some bookkeeping, but may be an
overall win.  We could also sort through the dp objects collected by the first
tree pass and identify dp->d_len objects larger than some threshold to be
excluded from parallel processing (as well as potentially memory hungry types
like dsp...)

Once all the dp objects are resolved, in the parent command we would do the
tree walk again and create our rt_db_internals for Manifold by reading the
facetized version of the geometry in from the working .g in lieu of the
original CSG object in the source .g file.  This would allow us to apply the
matrix and create the Manifold data structure, but on the BoT rather than the
original CSG data.