Start refactoring for different temp file approach
Untested as yet - just checkpointing a compiling state.  Instead of
writing out a large number of individual files, look at making one copy
of the original .g file to serve as the working template.  We will
identify unique leaf dp instances, and (in the working copy) overwrite
those solids with successful ged_tessellate output.  We can then perform
the tree walk to prepare Manifold inputs on the working copy, with the
BoT information in the correct place ahead of time.  Only the final
output of the booleval step gets written back to the original .g file -
all other activity will take place in the working copy.

This should also let us support resume without having to write anything
back to the original .g file - we can instead look for a previously
generated working file with the correct name, and have ged_tessellate
look there for state data to do things like avoiding known crashers if
we're resuming a previously incomplete process.  Hopefully the need to
do so will be much rarer - that's the whole point of ged_tessellate, to
isolate the parent controlling process from the risky stages - but we
still want to be able to recover a long conversion in that fashion if we
have to.