Pass multiple objects to the subprocess
Testing indicates launching the process is quite expensive, so we
definitely want to handle multiple objects per run.  To really do this
right we should probably implement a binary search if we encounter an
error, so we can do as many larger "batches" of objects as possible.

Windows is limited to 8191 characters on the command prompt, so that
will probably be the practical cap for how many objects we can pass in
on an argv.