Rough in a bisect method to zero in on bad inputs
If we're going to pass a large number of object into each tessellation
process, a single failure will mean that we don't know which one is the
poison pill.  Because we still want to bulk process as many as possible,
try a bisect that will re-try the conversion with half of each list.
Will keep splitting until individual bad objects are identified.

I suspect in a practical sense we should only do this bisecting for the
NMG method - the fallback methods are quite expensive in individual
processing times by comparison, and the process launching overhead will
almost certainly be noise for CM or SPSR.  Right now I'm testing havoc,
which doesn't get to those methods and has a large number of
fast-to-NMG-process CSG primitives, so we see big wins from large
processing lists.