Make a backup copy of the .g to use if we time out ged_tessellate
The problem with having to kill a runaway subprocess is that there's no way to
guarantee we don't do so at a point where the result is a corrupted, partially
written object (or even a corrupted .g file).  It's unlikely, but certainly not
impossible - accordingly, if we do have to hard stop we restore the pre-run
state of the working .g file.  This will mean potentially having to redo work
in subsequent bisects if we had a bunch of successful conversions before the
runaway, but as far as I can see that's the tradeoff we need to make in order
to allow both a) processing multiple objects in one subprocess to minimize
process setup/teardown overhead and b) allow timeout kills of infinite looping
processes.