Experiment with not closing output pipes in wait
After some hair pulling, checking the reason the child process was
failing appears to offer some insight on why we weren't getting the
return code.  Because the child was writing to the output pipe, and the
pipe got closed, the process got hit with a SIGPIPE.  If we leave the
output pipes open, we get the expected return code.

Doing this, on the other hand, breaks the abort test - need to look into
that.