Catch matrices that flip a BoT inside out
Goliath's SKTRACKdrivewheel2.c was producing inside out BoT instances
with its matrices - the primitives tessellated fine, but the matrix
turned them inside out.  These inside-out BoTs are a problem for
facetize in that they ARE technically manifold, so they do get processed
slowly and result in an "inside-out" component in the final mesh.  While
it's actually rather impressive this doesn't fail outright, it's also
not really the result we're looking for.

The impact of this "inside-out" matix can actually be seen in both
facetization and raytracing.  In the latter case, when a BoT is used for
one of the SKTRACKdrivewheel2.c primitives and the matrices are left
unchanged - there are no hits on that BoT.  A CSG primitive in the same
place DOES report a hit, so there's an inconsistency there that is
somewhat concerning, but that's a separate issue...

To catch and deal with this case for facetize, we check an exemplar
triangle face and intersecting ray segment that report an "outward
facing" first hit with the IDN matrix, and see if the result changes
after application of the currently active matrix in the tree walk.  If
it does, we reverse the BoT orientation for the purposes of the boolean
operation, essentially "undoing" the inside-out effect from the matrix.
This lets Goliath convert without SKTRACKdrivewheel2.c being inside-out.