Ignore combs with empty trees in the hierarchy
The way conversion.sh is set up right now, if we don't get a .bot object
out it's a FAIL.  However, for empty regions, that's actually the
expected outcome.  Can't just check for that though - there can be cases
like truck.g's g1 comb that have a tree full of regions with empty
trees.

Kinda wondering if the thing to do here is create an empty BoT as the
output for these cases, to distinguish between failure and "successfully
empty"...