Undo 'y' argv filter from 4b6850a671
Unfortunately, this filter breaks MGED command line processing.  The
following command illustrates the problem:

./bin/mged share/db/axis.g l y

It should print out the info for the 'y' TGC, but the filter strips out
the 'y' primitive name and it never reaches the 'l' routine.

We'll need to find another way to handle the "unrecognized command"
errors.