fix timeout handling
in the event of a timeout the subprocess is terminated and destroyed.
This sets the subprocess_join up for failure (atleast on Windows) as the
process handle was just forcefully closed. As a result, the join fails
and returns the function before the .g is restored. Reduce complexity by
handling the file restore and returning an error after killing the
timed-out process.

Also, modify to use clarifying returns.