Document a bug with TGC primitive tessellation
When one and only one of the c or d axis of a tgc is at or below the threshold
length of tol->dist/2, the rt_tgc_tess function fails to produce valid
output.  If both c and d are sufficiently small we succeed, but zeroing
out just one of them currently produces bad results.