Fully successful manifold cleanup of Generic_Twin
Adding a preprocessing step used by the geobox mesh repair feature in
geogram, we can also process the four components in Generic_Twin.g that
previously couldn't be handled.  With this change, we appear to be
successfully producing manifold outputs for ALL non-plate-mode BoT
objects in Generic_Twin that initially report as non-manifold per the
Manifold library's test.

Of course if enough of the original shape is missing, the results of
hole filling can't recover the original shape - but for a lot of "almost
closed" cases this looks as if it could be quite helpful.