Make in-place repair an option to bot manifold
This is for automatically repairing all bots in a file or hierarchy at
once via search exec in MGED:

search -type bot ! -type plate -exec "bot manifold -i {}" ";"

The above line will check all the non plate mode bots to see if they are
manifold.  If not, repair will be attempted to update the BoT to be
manifold via either small adjustments by the Manifold library or (if
that fails, which is more typical) hole filling using geogram.